pyproject.toml
[project]
name = "pega-pega"
version = "0.1.0"
description = "Multi-protocol request logger/catcher - catch and display all incoming requests"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "websockets>=12.0",
    "paramiko>=3.0",
    "cryptography>=41.0",
    "pyyaml>=6.0",
    "pyasn1>=0.5",
    "jinja2>=3.1",
]

[project.scripts]
pega-pega = "pega_pega.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
